Refactor debug_info.rs from debug/mod.rs
