+++ cli/src/debugger.rs
@@ -4,7 +4,7 @@ use capstone::Capstone;
 use num_traits::Num;
 use probe_rs::{
     architecture::arm::Dump,
-    debug::{registers::Registers, stack_frame::StackFrame, DebugInfo, VariableName},
+    debug::{debug_info::DebugInfo, registers::Registers, stack_frame::StackFrame, VariableName},
     Core, CoreRegisterAddress, MemoryInterface,
 };
 use std::fs::File;
