+++ probe-rs/tests/inlined_function.rs
@@ -1,10 +1,10 @@
-use probe_rs::debug::{DebugError, DebugInfo};
+use probe_rs::debug::{debug_info::DebugInfo, DebugError};
 
 type TestResult = Result<(), DebugError>;
 
 #[test]
 fn function_name_of_inlined_function_1() -> TestResult {
-    let di = DebugInfo::from_file("tests/inlined-function").unwrap();
+    let di = modname::DebugInfo::from_file("tests/inlined-function").unwrap();
 
     let address = 0x15e;
 
@@ -19,7 +19,7 @@ fn function_name_of_inlined_function_1() -> TestResult {
 
 #[test]
 fn name_of_function_containing_inlined_function_1() -> TestResult {
-    let di = DebugInfo::from_file("tests/inlined-function").unwrap();
+    let di = modname::DebugInfo::from_file("tests/inlined-function").unwrap();
 
     let address = 0x15e;
 
@@ -34,7 +34,7 @@ fn name_of_function_containing_inlined_function_1() -> TestResult {
 
 #[test]
 fn function_name_of_inlined_function_2() -> TestResult {
-    let di = DebugInfo::from_file("tests/inlined-function").unwrap();
+    let di = modname::DebugInfo::from_file("tests/inlined-function").unwrap();
 
     let address = 0x154;
 
@@ -49,7 +49,7 @@ fn function_name_of_inlined_function_2() -> TestResult {
 
 #[test]
 fn name_of_function_containing_inlined_function_2() -> TestResult {
-    let di = DebugInfo::from_file("tests/inlined-function").unwrap();
+    let di = modname::DebugInfo::from_file("tests/inlined-function").unwrap();
 
     let address = 0x154;
 
@@ -64,7 +64,7 @@ fn name_of_function_containing_inlined_function_2() -> TestResult {
 
 #[test]
 fn function_name_of_non_inlined_function() -> TestResult {
-    let di = DebugInfo::from_file("tests/inlined-function").unwrap();
+    let di = modname::DebugInfo::from_file("tests/inlined-function").unwrap();
 
     let address = 0xf4;
